extract authentication methods from protocol
